// SDVXDraw/ClientApp/src/utilities/enums.ts
export enum DifficultyType {
    NOVICE = 0,
    ADVANCED = 1,
    EXHAUST = 2,
    INFINITE = 3,
    MAXIMUM = 4
}

export enum InfVersion {
    INF = 2,
    GRV = 3,
    HVN = 4,
    VVD = 5
}

export enum GameVersion {
    BOOTH = 1,
    INFINITE_INFECTION = 2,
    GRAVITY_WARS = 3,
    HEAVENLY_HAVEN = 4,
    VIVIDWAVE = 5
}

export enum ChartStatus {
    NONE,
    PROTECTED,
    VETOED
}